Change grid container border, padding, background color
The change in the js file was because of the event bubbling which was causing a bug:
after the mouse entered the area of the grid container element or entered any of its children,
not only the background color of the child cells changed (which is the desired behavior), but
the grid container itself changed its background color.

This is an unfortunate side effect of using the "mouseover" event listener on the parent div, instead of
using hundreds or thousands of "mouseenter" event listeners on the child divs, which im willing to accept.