Fix shading related bug
If not fixed, the user can't color the shaded cells with a 100% bg color in normal mode.
meaning if he want to increase the opacity, his only way is to use shading; but he should
be able to color these half opacity cells in normal mode and make them 100% opacity directly
by one move.